+++ fedsim/__init__.py
@@ -3,15 +3,16 @@
 FedSim
 ------
 
- a simple Federated Learning simulator!
+Comprehensive and flexible Federated Learning Simulator!
 
 """
 
-__version__ = "0.5.0"
+__version__ = "0.6.0"
 
 from . import datasets
 from . import distributed
 from . import local
+from . import losses
 from . import lr_schedulers
 from . import models
 from . import scores
@@ -24,5 +25,6 @@ __all__ = [
     "local",
     "lr_schedulers",
     "scores",
+    "losses",
     "utils",
 ]