+++ fedsim_cli/fedsim_cli.py
@@ -7,6 +7,7 @@ import os
 
 import click
 
+from . import __version__
 from .fed_learn import fed_learn
 from .fed_tune import fed_tune
 
@@ -16,6 +17,7 @@ os.environ["LANG"] = "C.UTF-8"
 
 
 @click.group()
+@click.version_option(__version__)
 def cli():
     pass
 
@@ -24,5 +26,9 @@ cli.add_command(fed_learn)
 cli.add_command(fed_tune)
 
 
-if __name__ == "__main__":
+def main():
     cli()
+
+
+if __name__ == "__main__":
+    main()